Fix direct connection and other bugs

+++ src/main/java/it/simonedegiacomi/storage/handlers/ws/TrashHandler.java
@@ -19,6 +19,7 @@ import it.simonedegiacomi.sync.FileSystemWatcher;
 import org.apache.log4j.Logger;
 
 import java.io.IOException;
+import java.security.InvalidParameterException;
 import java.util.List;
 
 /**
@@ -28,19 +29,46 @@ import java.util.List;
  */
 public class TrashHandler {
 
+    /**
+     * Logger of the class
+     */
     private final static Logger log = Logger.getLogger(TrashHandler.class);
 
+    /**
+     * Gson
+     */
     private final Gson gson = MyGsonBuilder.create();
 
+    /**
+     * Path of the GoBox files folder
+     */
     private final String PATH = Config.getInstance().getProperty("path");
 
+    /**
+     * Event emitter
+     */
     private final EventEmitter emitter;
 
+    /**
+     * ile system watcher
+     */
     private final FileSystemWatcher watcher;
 
+    /**
+     * Database
+     */
     private final StorageDB db;
 
     public TrashHandler (StorageEnvironment env) {
+        if (env.getDB() == null)
+            throw new InvalidParameterException("Environment without db");
+
+        if (env.getEmitter() == null)
+            throw new InvalidParameterException("Environment without event emitter");
+
+        if (env.getSync() == null || env.getSync().getFileSystemWatcher() == null)
+            throw new InvalidParameterException("Environment without file system watcher");
+
         this.db = env.getDB();
         this.emitter = env.getEmitter();
         this.watcher = env.getSync().getFileSystemWatcher();
@@ -60,6 +88,7 @@ public class TrashHandler {
 
                 // Prepare the response
                 JsonObject response = new JsonObject();
+
                 if (!json.has("toTrash") || !json.has("file")) {
                     response.addProperty("success", false);
                     response.addProperty("error", "missing file and/or trash flag");
@@ -73,28 +102,37 @@ public class TrashHandler {
                     // Wrap the file from the request
                     GBFile file = gson.fromJson(json.get("file"), GBFile.class);
                     GBFile dbFile = db.getFile(file, true, false);
-                    dbFile.setPrefix(PATH);
 
-                    // Assert that the file exist
                     if (dbFile == null) {
                         response.addProperty("success", false);
                         response.addProperty("error", "file not found");
                         return response;
                     }
 
+                    dbFile.setPrefix(PATH);
+
                     // Hide the file
                     watcher.startIgnoring(dbFile.toFile());
+
                     dbFile.setTrashed(toTrash);
                     MyFileUtils.moveTrash(dbFile);
+
                     watcher.stopIgnoring(dbFile.toFile());
 
                     // Update the db
-                    db.trashFile(dbFile, toTrash);
+                    SyncEvent event = db.trashFile(dbFile, toTrash);
 
+                    // advise all the other clients
+                    emitter.emitEvent(event);
+
+                    // Complete the response
                     response.addProperty("success", true);
                 } catch (StorageException ex) {
+                    log.warn(ex.toString(), ex);
                     response.addProperty("success", false);
+                    response.addProperty("error", ex.toString());
                 } catch (IOException ex) {
+                    log.warn(ex.toString(), ex);
                     response.addProperty("success", false);
                     response.addProperty("error", ex.toString());
                 }
@@ -144,7 +182,6 @@ public class TrashHandler {
             @WSQuery(name = "removeFile")
             @Override
             public JsonElement onQuery (JsonElement data) {
-                JsonObject json = data.getAsJsonObject();
 
                 // Create the response object
                 JsonObject res = new JsonObject();
@@ -155,13 +192,26 @@ public class TrashHandler {
                     GBFile file = gson.fromJson(data, GBFile.class);
                     GBFile dbFile = db.getFile(file, true, true);
 
-                    // Tell the internal client to ignore this event
-                    watcher.startIgnoring(dbFile.toFile());
+                    if (dbFile == null) {
+                        res.addProperty("success", false);
+                        res.addProperty("error", "File not found");
+                        return res;
+                    }
+
+                    boolean alreadyTrashed = dbFile.isTrashed();
+                    if (!alreadyTrashed) {
+                        dbFile.setPrefix(PATH);
+
+                        // Tell the internal client to ignore this event
+                        watcher.startIgnoring(dbFile.toFile());
+                    }
 
                     // Remove the file from the file system
                     MyFileUtils.delete(dbFile);
 
-                    watcher.stopIgnoring(dbFile.toFile());
+                    if (!alreadyTrashed) {
+                        watcher.stopIgnoring(dbFile.toFile());
+                    }
 
                     // And then remove the file from the database
                     SyncEvent event = db.removeFile(dbFile);
@@ -206,14 +256,27 @@ public class TrashHandler {
                         db.findPath(file);
                         db.findChildren(file);
 
-                        // Set the environment prefix
-                        file.setPrefix(PATH);
+                        if (file.isTrashed()) {
+
+                            // Set the environment prefix
+                            file.setPrefix(PATH);
+
+                            // Ignore the event
+                            watcher.startIgnoring(file.toFile());
+                        }
 
                         // Delete the hidden file
                         MyFileUtils.delete(file);
 
+                        if (file.isTrashed()) {
+                            watcher.stopIgnoring(file.toFile());
+                        }
+
                         // Remove form the database
-                        db.removeFile(file);
+                        SyncEvent event = db.removeFile(file);
+
+                        // Advise the other clients
+                        emitter.emitEvent(event);
                     }
 
                     response.addProperty("success", true);
